add: test-add_const_data-bug
