+++ deploy/run-msvc.bat
@@ -18,7 +18,7 @@ rem ============================================================================
     set "order=msvc2017:64:debug:static"
     set "order=msvc2019:64:debug:static"
 
-    set "order=msvc2010:64:debug:static"
+    set "order=msvc2008:64:debug:static"
 
     call "%eDIR_BAT_ENGINE%\run.bat"  ^
         "--generate: cmake-makefiles" ^
